Streaming support
This commit modifies the /api/completion endpoint so that it streams
results back from the underlying LLM as JSON-Lines (line delimited JSON
objects). The first object contains information about the LLM request so
that the page can be updated with this information. Subsequent lines
contain objects that contain a partial response from the LLM.

The client then streams the results back and displays them as they
become available.